Refactor str package as a standard go package practice
