+++ utils/str/strings.go
@@ -1,4 +1,4 @@
-package utils
+package str
 
 import (
 	"errors"
